add .int() to z.number()

+++ packages/api/src/router/notif.ts
@@ -12,7 +12,7 @@ export const notifRouter = createTRPCRouter({
    *  @returns an array of message objects
    */
   getDMsFromChatId: publicProcedure
-    .input(z.object({ id: z.number() }))
+    .input(z.object({ id: z.number().int() }))
     .query(async ({ ctx, input }) => {
       const { prisma } = ctx
 
@@ -50,7 +50,7 @@ export const notifRouter = createTRPCRouter({
    *  @returns an array of objects containing the chatId, and an array containing the most recent message for that chat
    */
     getMessagePreviewsFromUserId: publicProcedure
-    .input(z.object({ id: z.number() }))
+    .input(z.object({ id: z.number().int() }))
     .query(async ({ ctx, input }) => {
       const { prisma } = ctx
 
@@ -103,7 +103,7 @@ export const notifRouter = createTRPCRouter({
       return finalMessageForEachDMChat;
     }),
 
-    
+
   /**
    *  @remarks
    *  This returns the miscellaneous notifications for a user.
@@ -112,7 +112,7 @@ export const notifRouter = createTRPCRouter({
    *  @returns an array of notification objects
    */
     getMiscNotifsFromUserId: publicProcedure
-    .input(z.object({ id: z.number() }))
+    .input(z.object({ id: z.number().int() }))
     .query(async ({ ctx, input }) => {
       const { prisma } = ctx
